backend/pyproject.toml
[tool.poetry]
name = "mensetu-backend"
version = "0.1.0"
description = "Interview training support backend (FastAPI + Supabase)"
authors = ["ootomonaiso <root@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
httpx = "^0.27.0"
supabase = "^2.5.3"
redis = "^5.0.1"
aiofiles = "^24.1.0"
python-multipart = "^0.0.9"
orjson = "^3.10.6"
celery = "^5.3.6"
faster-whisper = "^1.0.3"
librosa = "^0.10.2.post1"
google-generativeai = "^0.8.0"
tenacity = "^9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
ruff = "^0.6.4"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
